fix error introduced by 4f44bd41223d5732a076b46ad964214fb83418f9.
attachments must be fetched so they are compared against the cover
URL.

this is inneficient. something must be done about it in the future.
perhaps postgres can do this COVER IN ATTACHMENTS check instead of
our server, or we could do better cover image x attachment handling
so the check isn't needed.